this c/m is a/b work on notifs,ai.fronts.lay.acti.scree + subscr,draw_boq,whatsap.back.lay - stora.quota.add.prod.safety

+++ backend/app/modules/subscriptions/models.py
@@ -4,7 +4,7 @@ import uuid
 from datetime import datetime
 from decimal import Decimal
 from uuid import UUID
-from sqlalchemy import Boolean, DateTime, Enum, ForeignKey, Index, Integer, JSON, Numeric, String, Text, UniqueConstraint
+from sqlalchemy import BigInteger, Boolean, DateTime, Enum, ForeignKey, Index, JSON, Numeric, String, Text, UniqueConstraint
 from sqlalchemy.dialects.postgresql import UUID as PGUUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.core.database import Base
@@ -276,7 +276,7 @@ class UsageCounter(Base, TimestampMixin):
   )
 
   quantity: Mapped[int] = mapped_column(
-    Integer,
+    BigInteger,
     nullable=False,
     default=0,
   )
